requirements.txt
pyspark==3.5.0
databricks-connect==13.3.*
delta-spark==3.0.0
pandas==2.1.0
numpy==1.24.3
scikit-learn==1.3.0
mlflow==2.8.0
apache-airflow==2.7.1
python-dotenv==1.0.0
pytest==7.4.0
black==23.7.0
flake8==6.1.0 
eventregistry